Set ECB in-use flag after listen-for-packet call.
We still don't actually receive any packets yet, but setting the in-use
flag takes ownership of the ECBs. This is the minimal change needed to
get the vanilla-utilities version of IPXSETUP to run without crashing.
Without the in-use flag set, IPXSETUP thinks that a packet has been
received by the driver and performs an integer underflow when trying to
process the "received" packet.