Add implementation of trampoline function.
The redirector interrupt can return a function pointer that can be
called to access the IPX API instead of calling the interrupt. This can
presumably be more efficient since it does not involve triggering an
interrupt. Quake and the later version of Doom's IPXSETUP both use this
approach. However, for the first implementation here we use an
implementation that just invokes the interrupt, since it's easier than
writing a bunch of inline asm.