Allow TSR to be unloaded with `dali /u`.
This adds an additional hook into the redirector ISR that unloads
the hooked interrupt vectors, as opposed to previously where it was
called on exit. We must also do the same with the mTCP stack since
it hooks its own vectors. Also pass back the PSP segment of the
loaded driver so that the caller can invoke DOS to free the memory
that was being used (it is not safe for it to do this itself since
it would be freeing its own executing code).