Increase ISR stack size.
I believe this may fix some of the stability issues that people were
experiencing with previous versions. Although we switch stacks, we were
still only using a pretty small stack size, and the MTCP stack is
written in C++ which probably doesn't help. So bump it up to 4KiB.
On my DOS machine this resolves crashes when EMM386 is loaded.