Get rid of "IPX trampoline".
I put this in as a temporary hack until I figured out how to do the
redirector-based interruptless API calls. However, thinking this
through a bit, I think all that's really needed here is to return the
address of the normal ISR function. Since it's an "interrupt"
function it already takes care of saving and restoring registers, as
well as switching the data segment to Dali's segment. After testing,
this does indeed appear to work correctly.